Покращено структуру проекту
