+++ package.json
@@ -14,7 +14,8 @@
     "next": "14.2.7",
     "react": "^18",
     "react-dom": "^18",
-    "typo-js": "^1.2.4"
+    "typo-js": "^1.2.4",
+    "zustand": "^4.5.5"
   },
   "devDependencies": {
     "@types/node": "^20",
