+++ src/types/searchTypes.d.ts
@@ -1,0 +1,17 @@
+export interface Ware {
+	id: number;
+	shortName: string;
+	longName: string;
+	price: number;
+	tag: string;
+	discount: number;
+	deliveryOption: string;
+	storeAvailability: string;
+	rating: number;
+	imageSrc: string;
+}
+
+export interface Article {
+	title: string;
+	keywords: string[];
+}
